+++ tests/test_dag.py
@@ -19,10 +19,10 @@ from tiledb.cloud import client
 from tiledb.cloud import dag
 from tiledb.cloud import tasks
 from tiledb.cloud import tiledb_cloud_error as tce
+from tiledb.cloud._common import visitor
 from tiledb.cloud._results import decoders
 from tiledb.cloud._results import results
 from tiledb.cloud._results import stored_params as sp
-from tiledb.cloud._results import visitor
 from tiledb.cloud.dag import dag as dag_dag
 from tiledb.cloud.rest_api import models
