+++ tiledb/cloud/_common/json_safe.py
@@ -1,10 +1,14 @@
-from typing import Any
+from typing import Generic, TypeVar
 
 import attrs
 
+_T_co = TypeVar("_T_co", covariant=True)
 
-@attrs.define(frozen=True, slots=True)
-class Value:
+
+# This really should be a `slots` class, but setting `slots=True` breaks things
+# in python 3.6.
+@attrs.define(frozen=True, slots=False)
+class Value(Generic[_T_co]):
     """Sentinel for a value that is known to be JSON-serializable.
 
     In cases where you know we're generating JSON-safe values where the
@@ -13,4 +17,4 @@ class Value:
     you're generating into this.
     """
 
-    value: Any
+    value: _T_co
